FuelOptimizedRouteAPIProject/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
requests>=2.31.0
python-decouple>=3.8
